// src/users/dto/update-user.dto.ts
import { z } from 'zod';
import { UserRole } from '@prisma/client'; // Assuming UserRole is generated by Prisma

export const UpdateUserSchema = z.object({
  email: z.string().email({ message: 'Invalid email address' }).optional(),
  username: z.string().min(3, { message: 'Username must be at least 3 characters long' }).optional(),
  password: z.string().min(8, { message: 'Password must be at least 8 characters long' }).optional(),
  role: z.nativeEnum(UserRole).optional(),
});

export type UpdateUserDto = z.infer<typeof UpdateUserSchema>;